Streamline linear policy; spike improvements.
A large commit. Improvements to the linear RA policy include:
- reordering the clearing of RA register's tag and storing
for store instructions (prevents traps on stores from deleting tags)
- removed prohibition on loading RA's into other registers
(memcpy may do this whenever it wants, for example)

Improvements to spike include:
- the ability to watch a specific memory address and record the
PC of the last 1024 writes. use by starting spike with -t, and
watching an address with the watch command 'watch CORE ADDR'
and looking at the last few writes with 'when CORE NUMTOPRINT'
(NUMTOPRINT is 16 if not provided)
  - new test that can take advantage of this is in test/need_linux

Other improvements:
- linux/build_linux.sh now points to riscv-mit version of linux,
which fixes issues with sigreturn and the linear ra policy
- included new test for misaligned accesses (currently,
riscv-linux on spike generates sigbus correctly)