C++ exceptions work with modified libgcc_s.so.1.
This is a patched libgcc_s.so.1 (I replaced
an instruction using a hex editor) that allows exception
handling to work. If you want to use it, build your disk with
linux/cxx_exceptions/setup_disk.sh.
As an aside, we no longer restore the 11th floating point
register on an exception. Code that needs this and exceptions
won't work.