Minor changes to pc/reg/m/wreg/wregt.
Now, pc/reg/m/wreg/wregt commands no longer require
a processor number (if the command accepts an expression
as a series of arguments, the situation where a processor num
and an expression with multiple arguments are provided is
ambiguous, but we try to make it work).
Also removed a debug print.