Added tag_signals test for new Linux version.
New Linux kernel allows for tag traps to generate
SIGUSR2, and user programs to handle that signal.
test/single-file-tests/rop/tag_signals.c shows this off.
Note that generating SIGUSR2 from a tag violation and NOT
having a handler may lead to an infinite loop (because
the default action for SIGUSR2 is to be ignored).

Also removed binary file from test/single-file-tests/forward-edge.